requirements.txt
streamlit>=1.32.0
pdfplumber>=0.10.3
python-docx>=1.1.0
google-generativeai>=0.5.0
scikit-learn>=1.4.0
plotly>=5.19.0
pandas>=2.2.0
numpy>=1.26.0
reportlab>=4.1.0
SQLAlchemy>=2.0.0
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
Pillow>=10.2.0
wordcloud>=1.9.3
spacy>=3.7.0
nltk>=3.8.1
